refactor: consolidate wave height formatting (3 files → 1)
Merge wave-height-formatter.ts and format-utils.ts into wave-formatters.ts
and surf-data.ts. Rename three formatWaveHeight variants to unique names:
formatWaveHeightDecimal, formatWaveHeightRange, formatWaveHeightBucket.
Update ~17 consumer imports across the codebase.

Phase 1A of codebase redundancy cleanup.
