+++ components/session-forms/SessionDetailsSection.tsx
@@ -16,7 +16,7 @@ import {
   Loader2,
 } from "lucide-react";
 import { useSessionForecast } from "@/hooks/use-session-forecast";
-import { formatWaveHeightRangeString } from "@/lib/utils/wave-height-formatter";
+import { formatWaveHeightRangeString } from "@/lib/utils/wave-formatters";
 import { SET_WAVE_VARIANCE } from "@/lib/utils/wave-height-transformer";
 import { SimpleCardLayout } from "@/components/ui/form-layout";
 import { Input } from "@/components/ui/input";
