+++ lib/utils/surf-call-logic.ts
@@ -11,7 +11,7 @@ import type { Beach } from '@/types/database';
 import type { PersonalizedForecastWindow } from '@/types/personalization';
 import type { EnhancedForecastEntity } from '@/types/forecast';
 import { computeTrendTags, type TrendTag } from '@/lib/scoring';
-import { formatTideHeight, formatWaveHeight } from '@/lib/formatters/surf-data';
+import { formatTideHeight, formatWaveHeightRange as formatWaveHeight } from '@/lib/formatters/surf-data';
 import { parseSkillLevel, SKILL_WAVE_RANGES } from '@/lib/domains/user-preferences/skill-level';
 
 // ============================================================================
